test: Add test cases to common state flow #14
